포지션 값 추가

+++ src/main/kotlin/net/prayforyou/backend/application/GetBattlePositionService.kt
@@ -12,7 +12,7 @@ class GetBattlePositionService(
     private val battlePositionProvider: BattlePositionProvider
 ) {
 
-    fun getBattlePositionByXandY(x: Int, y: Int): BattlePosition {
+    fun getBattlePositionByXandY(x: Double, y: Double): BattlePosition {
         val battlePositions = battlePositionProvider.findAll()
         return try {
             battlePositions.first { it.isContainsBattlePosition(x, y) }
